Conflicts:
	components/nexus-compat/pom.xml
	components/nexus-compat/src/main/java/org/sonatype/nexus/compat/EventBusApplicationEventMulticaster.java
	components/nexus-core/src/main/java/org/sonatype/guice/nexus/scanners/NexusTypeCache.java
	components/nexus-core/src/main/java/org/sonatype/guice/nexus/scanners/NexusTypeVisitor.java
	components/nexus-core/src/main/java/org/sonatype/nexus/error/reporting/SetProxyPropertiesInspector.java
	components/nexus-core/src/main/java/org/sonatype/nexus/events/DebugEventInspector.java
	components/nexus-core/src/main/java/org/sonatype/nexus/events/DefaultEventInspectorHost.java
	components/nexus-core/src/main/java/org/sonatype/nexus/index/events/ReindexRepositoriesEvent.java
	components/nexus-core/src/main/java/org/sonatype/nexus/index/events/ReindexRepositoriesRequest.java
	components/nexus-core/src/main/java/org/sonatype/nexus/proxy/attributes/StorageFileItemInspector.java
	components/nexus-core/src/main/java/org/sonatype/nexus/proxy/events/AsynchronousEventInspectorWrapper.java
	components/nexus-core/src/main/java/org/sonatype/nexus/proxy/events/EventInspector.java
	components/nexus-core/src/main/java/org/sonatype/nexus/scheduling/NexusTaskFailureErrorManagerReporter.java
	components/nexus-core/src/main/java/org/sonatype/nexus/util/StringDigester.java
	components/nexus-core/src/test/java/org/sonatype/nexus/events/DefaultEventInspectorHostTest.java
	components/nexus-core/src/test/java/org/sonatype/nexus/proxy/AbstractProxyTestEnvironment.java
	plugins/internal/nexus-wonderland-plugin/src/main/resources/static/js/Nexus/wonderland/Icons.js
	plugins/internal/nexus-wonderland-plugin/src/main/resources/static/js/nexus-wonderland-plugin-boot.js
	plugins/maven/nexus-archetype-plugin/src/main/java/org/sonatype/nexus/plugins/mac/MacPluginEventSubscriber.java

+++ components/nexus-core/src/main/java/org/sonatype/nexus/proxy/attributes/DefaultAttributesHandler.java
@@ -18,6 +18,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.nio.file.Files;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -35,15 +36,18 @@ import org.sonatype.nexus.proxy.item.RepositoryItemUid;
 import org.sonatype.nexus.proxy.item.StorageCollectionItem;
 import org.sonatype.nexus.proxy.item.StorageFileItem;
 import org.sonatype.nexus.proxy.item.StorageItem;
+import org.sonatype.nexus.proxy.item.uid.IsMetadataMaintainedAttribute;
 import org.sonatype.nexus.proxy.repository.HostedRepository;
 import org.sonatype.nexus.proxy.repository.ProxyRepository;
 import org.sonatype.nexus.proxy.repository.Repository;
 import org.sonatype.nexus.proxy.repository.RepositoryKind;
 import org.sonatype.nexus.util.SystemPropertiesHelper;
+import org.sonatype.nexus.util.io.StreamSupport;
 
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.Lists;
 
-import org.codehaus.plexus.util.IOUtil;
+import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * The default implementation of AttributesHandler. Does not have any assumption regarding actual AttributeStorage it
@@ -94,29 +98,30 @@ public class DefaultAttributesHandler
       LAST_REQUEST_ATTRIBUTE_ENABLED);
 
   /**
-   * The actual value of lastRequest attribute's resolution.
+   * The actual value of lastRequest attribute's resolution. Note: is not final due to UT access, see
+   * setter method that is visible for testing.
    */
   private long lastRequestedResolution = LAST_REQUESTED_ATTRIBUTE_RESOLUTION;
 
   /**
    * The application configuration.
    */
-  private ApplicationConfiguration applicationConfiguration;
+  private final ApplicationConfiguration applicationConfiguration;
 
   /**
    * The attribute storage.
    */
-  private DelegatingAttributeStorage attributeStorage;
+  private final AttributeStorage attributeStorage;
 
   /**
    * The item inspector list.
    */
-  protected List<StorageItemInspector> itemInspectorList;
+  private final List<StorageItemInspector> itemInspectorList;
 
   /**
    * The item inspector list.
    */
-  protected List<StorageFileItemInspector> fileItemInspectorList;
+  private final List<StorageFileItemInspector> fileItemInspectorList;
 
   @Inject
   public DefaultAttributesHandler(ApplicationConfiguration applicationConfiguration,
@@ -125,7 +130,7 @@ public class DefaultAttributesHandler
                                   List<StorageItemInspector> itemInspectorList,
                                   List<StorageFileItemInspector> fileItemInspectorList)
   {
-    this.applicationConfiguration = applicationConfiguration;
+    this.applicationConfiguration = checkNotNull(applicationConfiguration);
 
     // do we need to waste CPU cycles at "transitioning" at all? Should not, ie, for new instances
     if (legacyAttributeStorage != null
@@ -141,8 +146,8 @@ public class DefaultAttributesHandler
     else {
       this.attributeStorage = new DelegatingAttributeStorage(attributeStorage);
     }
-    this.itemInspectorList = itemInspectorList;
-    this.fileItemInspectorList = fileItemInspectorList;
+    this.itemInspectorList = checkNotNull(itemInspectorList);
+    this.fileItemInspectorList = checkNotNull(fileItemInspectorList);
   }
 
   // ==
@@ -152,7 +157,7 @@ public class DefaultAttributesHandler
    *
    * @return the attribute storage
    */
-  public DelegatingAttributeStorage getAttributeStorage() {
+  public AttributeStorage getAttributeStorage() {
     return attributeStorage;
   }
 
@@ -166,30 +171,12 @@ public class DefaultAttributesHandler
   }
 
   /**
-   * Sets the item inspector list.
-   *
-   * @param itemInspectorList the new item inspector list
-   */
-  public void setItemInspectorList(List<StorageItemInspector> itemInspectorList) {
-    this.itemInspectorList = itemInspectorList;
-  }
-
-  /**
    * Gets the file item inspector list.
    *
    * @return the file item inspector list
    */
   public List<StorageFileItemInspector> getFileItemInspectorList() {
     return fileItemInspectorList;
-  }
-
-  /**
-   * Sets the file item inspector list.
-   *
-   * @param fileItemInspectorList the new file item inspector list
-   */
-  public void setFileItemInspectorList(List<StorageFileItemInspector> fileItemInspectorList) {
-    this.fileItemInspectorList = fileItemInspectorList;
   }
 
   // ======================================================================
@@ -199,15 +186,19 @@ public class DefaultAttributesHandler
   public boolean deleteAttributes(final RepositoryItemUid uid)
       throws IOException
   {
-    return getAttributeStorage().deleteAttributes(uid);
+    if (!isMetadataMaintained(uid)) {
+      return false;
+    }
+    else {
+      return getAttributeStorage().deleteAttributes(uid);
+    }
   }
 
   @Override
   public void fetchAttributes(final StorageItem item)
       throws IOException
   {
-    if (item instanceof StorageCollectionItem) {
-      // not storing attributes of directories
+    if (!isMetadataMaintained(item)) {
       return;
     }
 
@@ -218,13 +209,10 @@ public class DefaultAttributesHandler
     }
     else {
       // we are fixing md if we can
-
       ContentLocator is = null;
-
-      if (StorageFileItem.class.isAssignableFrom(item.getClass())) {
-        if (((StorageFileItem) item).getContentLocator().isReusable()) {
+      if (item instanceof StorageFileItem &&
+        ((StorageFileItem) item).getContentLocator().isReusable()) {
           is = ((StorageFileItem) item).getContentLocator();
-        }
       }
 
       storeAttributes(item, is);
@@ -235,8 +223,7 @@ public class DefaultAttributesHandler
   public void storeAttributes(final StorageItem item)
       throws IOException
   {
-    if (item instanceof StorageCollectionItem) {
-      // not storing attributes of directories
+    if (!isMetadataMaintained(item)) {
       return;
     }
 
@@ -247,8 +234,7 @@ public class DefaultAttributesHandler
   public void storeAttributes(final StorageItem item, final ContentLocator content)
       throws IOException
   {
-    if (item instanceof StorageCollectionItem) {
-      // not storing attributes of directories
+    if (!isMetadataMaintained(item)) {
       return;
     }
 
@@ -275,8 +261,7 @@ public class DefaultAttributesHandler
   public void touchItemCheckedRemotely(final long timestamp, final StorageItem storageItem)
       throws IOException
   {
-    if (storageItem instanceof StorageCollectionItem) {
-      // not storing attributes of directories
+    if (!isMetadataMaintained(storageItem)) {
       return;
     }
 
@@ -297,8 +282,7 @@ public class DefaultAttributesHandler
   public void touchItemLastRequested(final long timestamp, final StorageItem storageItem)
       throws IOException
   {
-    if (storageItem instanceof StorageCollectionItem) {
-      // not storing attributes of directories
+    if (!isMetadataMaintained(storageItem)) {
       return;
     }
 
@@ -306,7 +290,32 @@ public class DefaultAttributesHandler
         storageItem.getRepositoryItemAttributes());
   }
 
-  // ==
+  // ======================================================================
+  // Internal
+
+  protected boolean isMetadataMaintained(final StorageItem item) {
+    if (item instanceof StorageCollectionItem) {
+      // not storing attributes of directories
+      return false;
+    } else if (item.isVirtual()) {
+      // virtual items have no attributes (nor UID for that matter)
+      return false;
+    }
+    else {
+      return isMetadataMaintained(item.getRepositoryItemUid());
+    }
+  }
+
+  protected boolean isMetadataMaintained(final RepositoryItemUid uid) {
+    final Boolean isMetadataMaintained = uid.getAttributeValue(IsMetadataMaintainedAttribute.class);
+    if (isMetadataMaintained != null) {
+      return isMetadataMaintained.booleanValue();
+    }
+    else {
+      // safest
+      return true;
+    }
+  }
 
   protected void touchItemLastRequested(final long timestamp, final ResourceStoreRequest request,
                                         final RepositoryItemUid uid, final Attributes attributes)
@@ -361,119 +370,107 @@ public class DefaultAttributesHandler
     return doTouch;
   }
 
-  // ======================================================================
-  // Internal
-
   /**
-   * Expand custom item attributes using registered StorageFileItemInspector (for files) or StorageItemInspector (for
-   * everything else) components.
+   * Method that expands core (and custom if custom inspector exists, like provided by a plugin) item attributes on
+   * files having attributes maintained. This method may be called ONLY if {@link #isMetadataMaintained(StorageItem)}
+   * returns {@code true} and the content locator passed in is reusable. Currently this method is invoked only from
+   * {@link #storeAttributes(StorageItem, ContentLocator)} method.
    *
    * @param item    the item
-   * @param content the input stream
+   * @param content the reusable content locator
    */
-  protected void expandCustomItemAttributes(StorageItem item, ContentLocator content) {
-    // gather inspectors willing to participate first, to save file copying below
-    ArrayList<StorageFileItemInspector> handlingInspectors = new ArrayList<StorageFileItemInspector>();
-    for (StorageFileItemInspector inspector : getFileItemInspectorList()) {
-      if (inspector.isHandled(item)) {
-        handlingInspectors.add(inspector);
+  protected void expandCustomItemAttributes(final StorageItem item, final ContentLocator content) {
+    // Step1: deprecated stuff
+    // gather deprecated inspectors willing to participate first, to save file mucking below if possible
+    if (item instanceof StorageFileItem) {
+      final StorageFileItem fItem = (StorageFileItem) item;
+
+      final ArrayList<StorageFileItemInspector> handlingFileInspectors = Lists.newArrayList();
+      for (StorageFileItemInspector inspector : getFileItemInspectorList()) {
+        if (inspector.isHandled(fItem)) {
+          handlingFileInspectors.add(inspector);
+        }
       }
-    }
+      if (!handlingFileInspectors.isEmpty()) {
+        // everything in this if should NEVER happen, unless some
+        // 3rd party plugin uses StorageFileItemInspector for something
+        // as core has no more these
+        // We have to have a java.io.File at any cause, as deprecated StorageFileItemInspector needs one
+        // even if the ContentLocator we got is reusable
+        getLogger().warn(
+            "Deprecated StorageFileItemInspector components used: {}. Change your plugin to use StorageItemInspector instead!",
+            handlingFileInspectors);
 
-    if (handlingInspectors.isEmpty()) {
-      return;
-    }
+        boolean deleteTmpFile = false;
+        File tmpFile = null;
+        if (content != null) {
+          if (content instanceof FileContentLocator) {
+            deleteTmpFile = false;
+            tmpFile = ((FileContentLocator) content).getFile();
+          }
+          else {
+            getLogger().info(
+                "Doing a temporary copy of the \"{}\" item's content for expanding custom attributes. This should NOT happen, but is left in as \"fallback\"!",
+                item.getPath());
+            deleteTmpFile = true;
+            try {
+              tmpFile =
+                  File.createTempFile("px-" + fItem.getName(), ".tmp",
+                      applicationConfiguration.getTemporaryDirectory());
 
-    boolean deleteTmpFile = false;
-    File tmpFile = null;
+              try (final InputStream in = content.getContent(); final OutputStream out = new FileOutputStream(
+                  tmpFile)) {
+                StreamSupport.copy(in, out);
+                out.flush();
+              }
+            }
+            catch (IOException ex) {
+              getLogger().warn("Could not create file from {}", item.getRepositoryItemUid(), ex);
+              tmpFile = null;
+            }
+          }
+        }
 
-    if (content != null) {
-      if (content instanceof FileContentLocator) {
-        deleteTmpFile = false;
-        tmpFile = ((FileContentLocator) content).getFile();
-      }
-      else {
-        getLogger().info(
-            "Doing a temporary copy of the \""
-                + item.getPath()
-                +
-                "\" item's content for expanding custom attributes. This should NOT happen, but is left in as \"fallback\"!");
-
-        deleteTmpFile = true;
-
-        try {
-          InputStream inputStream = null;
-          OutputStream tmpFileStream = null;
-
+        if (tmpFile != null) {
           try {
-            // unpack the file
-            tmpFile =
-                File.createTempFile("px-" + item.getName(), ".tmp",
-                    applicationConfiguration.getTemporaryDirectory());
-
-            inputStream = content.getContent();
-
-            tmpFileStream = new FileOutputStream(tmpFile);
-
-            IOUtil.copy(inputStream, tmpFileStream);
-
-            tmpFileStream.flush();
-
-            tmpFileStream.close();
-          }
-          finally {
-            IOUtil.close(inputStream);
-
-            IOUtil.close(tmpFileStream);
-          }
-        }
-        catch (IOException ex) {
-          getLogger().warn("Could not create file from " + item.getRepositoryItemUid(), ex);
-
-          tmpFile = null;
-        }
-      }
-    }
-
-    if (StorageFileItem.class.isAssignableFrom(item.getClass())) {
-      StorageFileItem fItem = (StorageFileItem) item;
-
-      if (!fItem.isVirtual() && tmpFile != null) {
-        try {
-          // we should prepare a file for inspectors
-          for (StorageFileItemInspector inspector : handlingInspectors) {
-            if (inspector.isHandled(item)) {
+            for (StorageFileItemInspector inspector : handlingFileInspectors) {
+              // isHandled already invoked way above, so handlingFileInspectors has
+              // inspectors already stated they do want to handle this item
               try {
                 inspector.processStorageFileItem(fItem, tmpFile);
               }
               catch (Exception ex) {
                 getLogger().warn(
-                    "Inspector " + inspector.getClass() + " throw exception during inspection of "
-                        + item.getRepositoryItemUid() + ", continuing...", ex);
+                    "Inspector {} throw exception during inspection of {}, continuing...", inspector.getClass(),
+                    item.getRepositoryItemUid(), ex);
+              }
+            }
+          }
+          finally {
+            if (deleteTmpFile && tmpFile != null) {
+              try {
+                Files.delete(tmpFile.toPath());
+              }
+              catch (IOException e) {
+                tmpFile.deleteOnExit();
+                getLogger().warn("Could not delete tmp file!", e);
               }
             }
           }
         }
-        finally {
-          if (deleteTmpFile && tmpFile != null) {
-            tmpFile.delete();
-          }
-
-          tmpFile = null;
-        }
       }
     }
-    else {
-      for (StorageItemInspector inspector : getItemInspectorList()) {
-        if (inspector.isHandled(item)) {
-          try {
-            inspector.processStorageItem(item);
-          }
-          catch (Exception ex) {
-            getLogger().warn(
-                "Inspector " + inspector.getClass() + " throw exception during inspection of "
-                    + item.getRepositoryItemUid() + ", continuing...", ex);
-          }
+
+    // make it handled by "plain" StorageItemInspectors too
+    for (StorageItemInspector inspector : getItemInspectorList()) {
+      if (inspector.isHandled(item)) {
+        try {
+          inspector.processStorageItem(item);
+        }
+        catch (Exception ex) {
+          getLogger().warn(
+              "Inspector {} throw exception during inspection of {}, continuing...", inspector.getClass(),
+              item.getRepositoryItemUid(), ex);
         }
       }
     }
@@ -495,13 +492,5 @@ public class DefaultAttributesHandler
   @VisibleForTesting
   void setLastRequestedResolution(long lastRequestedResolution) {
     this.lastRequestedResolution = lastRequestedResolution;
-  }
-
-  /**
-   * For UT access!
-   */
-  @VisibleForTesting
-  void setAttributeStorage(final AttributeStorage attributeStorage) {
-    this.attributeStorage = new DelegatingAttributeStorage(attributeStorage);
   }
 }
